requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
Flask-WTF==1.1.1
Flask-Login==0.6.2
PyMySQL==1.1.0
Werkzeug==2.3.6
numpy==1.26.4
pandas==2.2.2
openpyxl==3.1.2
python-dotenv==1.0.0
email-validator==2.0.0.post2